requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
iniconfig==1.1.1
mccabe==0.7.0
packaging==22.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.7
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
typing-extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0